python-agents/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
langgraph>=0.0.1
openai>=1.0.0
supabase>=2.0.0
serpapi>=0.1.5
python-dotenv>=1.0.0

